some updates
